--- webclient/app/src/app/commons/appTemplateSelectCard/AppTemplateSelectCardStyles.js
@@ -1,11 +1,0 @@
-import {makeStyles} from "@mui/styles";
-
-const AppTemplateSelectCardStyles = makeStyles(theme => ({
-    actionsWrapper: {
-        justifyContent: "flex-end"
-    },
-    cardMedia: {
-        backgroundColor: "#000"
-    }
-}))
-export default AppTemplateSelectCardStyles;
