+++ webclient/app/src/app/features/appTemplates/appTemplateCard/AppTemplateCardStyles.js
