+++ webclient/app/src/app/features/apps/appTemplateSelectCard/AppTemplateSelectCard.jsx
