+++ webclient/app/src/app/features/apps/entities/ErSectionStyles.js
