+++ webclient/app/src/app/features/apps/entities/entityCard/EntityCardStyles.js
