+++ webclient/app/src/app/features/apps/entities/entityDialog/EntityDialogStyles.js
