+++ webclient/app/src/app/features/apps/entities/fieldAccordion/FieldAccordionStyles.js
