+++ webclient/app/src/app/pages/MainContentRouter.jsx
@@ -1,18 +1,18 @@
 import React from "react";
 import {Route} from "react-router-dom";
 
-import Home from "./home/Home";
-import AppOverview from "./appOverview/AppOverview";
-import AppEditor from "./appEditor/AppEditor";
-import AppTemplateOverview from "./appTemplate/AppTemplateOverview";
+import Home from "./apps/AppOverview";
+import AppOverview from "./apps/AppSummary";
+import AppEditor from "./apps/AppEditor";
+import AppTemplateOverview from "./appTemplates/AppTemplateOverview";
 
 function MainContentRouter() {
     return (
         <>
-            <Route path={"/app/:appId/edit"} component={AppEditor}/>
-            <Route exact path={"/app/:appId"} component={AppOverview}/>
+            <Route path={"/apps/:appId/edit"} component={AppEditor}/>
+            <Route exact path={"/apps/:appId"} component={AppOverview}/>
             <Route exact path={"/"} component={Home}/>
-            <Route path={"/apptemplate"} component={AppTemplateOverview}/>
+            <Route path={"/apptemplates/"} component={AppTemplateOverview}/>
             <Route path='/logout' component={() => {                
                 window.location.href = window.location.pathname + "api/user/logout"; 
                 return null;
