--- webclient/app/src/app/pages/appEditor/sections/conclusion/ConclusionSectionStyles.js
@@ -1,20 +1,0 @@
-import {makeStyles} from "@mui/styles";
-
-const ConclusionSectionStyles = makeStyles(theme => ({
-    downloadGrid: {
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-        width: "100%"
-    },
-    list: {
-        width: "100%"
-    },
-    root: {
-        paddingTop: "2rem"
-    },
-    listItemText: {
-        fontWeight: "bold"
-    }
-}));
-export default ConclusionSectionStyles;
