--- webclient/app/src/app/pages/appEditor/sections/generalSection/GeneralSectionStyles.js
@@ -1,8 +1,0 @@
-import {makeStyles} from "@mui/styles";
-
-const GeneralSectionStyles = makeStyles(theme => ({
-    padding: {
-        paddingBottom: "2rem"
-    }
-}));
-export default GeneralSectionStyles
