+++ webclient/app/src/app/pages/appTemplates/AppTempateOverviewStyles.js
