+++ webclient/app/src/app/pages/apps/AppOverviewStyles.js
@@ -1,0 +1,10 @@
+import {makeStyles} from "@mui/styles";
+
+const AppOverviewStyles = makeStyles(theme => ({
+    addFab: {
+        position: "absolute",
+        bottom: "5%",
+        right: "2%"
+    },
+}));
+export default AppOverviewStyles;
