--- webclient/app/src/app/pages/home/HomeStyles.js
@@ -1,10 +1,0 @@
-import {makeStyles} from "@mui/styles";
-
-const HomeStyles = makeStyles(theme => ({
-    addFab: {
-        position: "absolute",
-        bottom: "5%",
-        right: "2%"
-    },
-}));
-export default HomeStyles;
